Adding triple update at the end of atom resolving
