+++ .sqlx/query-859eaff9f915022c2cdbc34e270d6239c1ee9a0aa87a8de07d77edad0ea45403.json
@@ -1,0 +1,14 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "\n            UPDATE triple \n            SET label = subquery.new_label\n            FROM (\n                SELECT t.id,\n                    CASE \n                        WHEN t.subject_id = $1 THEN s.label\n                        WHEN t.predicate_id = $1 THEN p.label\n                        WHEN t.object_id = $1 THEN o.label\n                    END as new_label\n                FROM triple t\n                LEFT JOIN triple s ON t.subject_id = s.id\n                LEFT JOIN triple p ON t.predicate_id = p.id\n                LEFT JOIN triple o ON t.object_id = o.id\n                WHERE t.subject_id = $1 OR t.predicate_id = $1 OR t.object_id = $1\n            ) as subquery\n            WHERE triple.id = subquery.id",
+  "describe": {
+    "columns": [],
+    "parameters": {
+      "Left": [
+        "Numeric"
+      ]
+    },
+    "nullable": []
+  },
+  "hash": "859eaff9f915022c2cdbc34e270d6239c1ee9a0aa87a8de07d77edad0ea45403"
+}
