+++ consumer/src/mode/decoded/atom/atom_creation.rs
@@ -5,7 +5,8 @@ use crate::{
             atom::atom_supported_types::get_supported_atom_metadata,
             utils::{get_or_create_account, short_id},
         },
-        types::{DecodedConsumerContext, ResolveAtom},
+        resolver::types::ResolveAtom,
+        types::DecodedConsumerContext,
     },
     schemas::types::DecodedMessage,
     EthMultiVault::AtomCreated,
