+++ consumer/src/mode/resolver/mod.rs
@@ -1,3 +1,4 @@
 pub mod atom_resolver;
 pub mod ens_resolver;
 pub mod ipfs_resolver;
+pub mod types;
